Change: Massive change for everything to use the FSO (File System Object) struct for storing information about files and directories - this seems less efficient in some ways, but increases consistency.
Future:
I need to go through and tweak where the FSO are being copied versus accessed or moved (try to avoid copying data as much as possible). Add ability to do cmd line parameters. Ultimately turn this in to a shared library (figure out GPL 2/3 to include 'wolfcrypt' with source).